rewrote the setAuthToken, changed it to AuthTokenSet and effect the changes all over the work file.
