+++ src/components/DeleteEmployeeModal.jsx
@@ -1,5 +1,4 @@
-/* eslint-disable react/prop-types */
-// eslint-disable-next-line no-unused-vars
+
 import React from 'react'
 import {BiUserX} from 'react-icons/bi'
 import {ImCancelCircle} from 'react-icons/im'
