+++ src/components/HistoryPageNav.jsx
@@ -1,4 +1,3 @@
-// eslint-disable-next-line no-unused-vars
 import React from "react";
 import { NavLink } from "react-router-dom";
 import "../assets/styles/historyPage.css";
