+++ src/components/MainContentContainer.jsx
@@ -1,10 +1,9 @@
-/* eslint-disable react/prop-types */
 import {useState} from 'react';
 import Aside from "./Aside"
 import MainContainer from "./MainContainer"
-import SuccessSlip from "./SuccessModal/SuccessSlip"
+// import SuccessSlip from "./SuccessModal/SuccessSlip"
 import SuccessReceipt from "./SuccessModal/SuccessReceipt"
-
+import SuccessSlipp from './SuccessModal/SuccessSlipp';
 import NavPay from "./NavPay"
 
 function MainContentContainer({data,  sideBarIsOpen, setSideBarIsOpen }) {
@@ -39,7 +38,7 @@ function MainContentContainer({data,  sideBarIsOpen, setSideBarIsOpen }) {
           </div>
             <div className="w-screen md:w-[84%] md:relative h-full z-0 md:z-10 px-4">
                 <MainContainer data={data} togglePaySLip={togglePaySLip} />
-                <SuccessSlip showReceipt={showReceipt} popSlip={popSlip} togglePaySLip={togglePaySLip} handlePayNow={handlePayNow}/> 
+                <SuccessSlipp showReceipt={showReceipt} popSlip={popSlip} togglePaySLip={togglePaySLip} handlePayNow={handlePayNow}/> 
                 <SuccessReceipt showReceipt={showReceipt} closePopReceipt={closePopReceipt} />
             </div>
         </div>
