+++ src/components/SuccessModal/SuccessReceipt.jsx
@@ -8,7 +8,6 @@ import good from "../../assets/good.svg"
 import printer from "../../assets/printer.svg"
 import { displaydate } from "../date"
 import FetchTotalnetpay from "../FetchTotalnetpay"
-// import { setAuthToken } from "../setAuthToken";
 import AuthTokenSet from "../setAuthToken";
 import { useRef } from 'react';
 import { useReactToPrint } from 'react-to-print';
